Increase max test time for melt_pool_fluid.fluid (#6)
Debug testing on CIVET leads to a test time of ~413s, but this is not reproducible on Mac or Linux testing (via rod and moosebuild), which clears the standard 300s limit.

Increasing the max test time seemed preferable over changing the test to HEAVY, as there are only a few tests checking core capability. So, this should continue to be run on each standard recipe.